requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
eval_type_backport==0.2.0
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lightning==2.4.0
lightning-utilities==0.11.9
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
psutil==6.1.0
py-cpuinfo==9.0.0
pycocotools==2.0.8
pydantic==2.10.2
pydantic_core==2.27.1
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
simsimd==6.2.1
six==1.16.0
stringzilla==3.11.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchinfo==1.8.0
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.40
ultralytics-thop==2.0.12
urllib3==2.2.3
yarl==1.18.3
